Eslint . fix

+++ eslint.config.js
@@ -36,7 +36,7 @@ export default [
     rules: {
       indent: ["error", 2],
       "linebreak-style": ["error", "unix"],
-      quotes: ["error", "single"],
+      quotes: ["error", "double"],
       semi: ["error", "always"],
       "no-unused-vars": ["warn"],
       "no-console": ["warn", { allow: ["warn", "error"] }],
